// frontend/src/generated/api.ts
export interface ApiSchemas {
// ...
  "SolveRequest": { "deformation_scale"?: number; "distributed_loads"?: Array<ApiSchemas["DistributedLoadInput"]>; "elements": Array<ApiSchemas["reused_cores__frame3d_linear__api__ElementInput"]>; "include_plots"?: boolean; "nodal_loads"?: Array<ApiSchemas["NodalLoadInput"]>; "nodes": Array<ApiSchemas["reused_cores__frame3d_linear__api__NodeInput"]>; "number_of_points"?: number; "plot_dpi"?: number; "section_points"?: Array<ApiSchemas["SectionPointInput"]>; "supports"?: Array<ApiSchemas["SupportInput"]> }
  "SolveResponse": { "active_dof_count": number; "displacement_dof_order": string; "elements": Array<ApiSchemas["ElementResultOutput"]>; "field_recovery": string; "force_dof_order": string; "free_dofs": Array<number>; "inactive_dof_vectors": Array<Array<number>>; "nodal_displacements": Array<ApiSchemas["NodalDisplacementOutput"]>; "nodal_reactions": Array<ApiSchemas["NodalReactionOutput"]>; "plots": ({ [key: string]: ApiSchemas["PlotOutput"] }) | (null); "restrained_dofs": Array<number>; "section_force_convention": string; "section_points": Array<ApiSchemas["SectionPointInput"]>; "validation": ApiSchemas["GlobalValidationOutput"] }
// ...
  "SupportInput": { "axes"?: ([[number, number, number], [number, number, number], [number, number, number]]) | (null); "node_id": number; "rx"?: boolean; "rx_value"?: number; "ry"?: boolean; "ry_value"?: number; "rz"?: boolean; "rz_value"?: number; "u"?: boolean; "u_value"?: number; "v"?: boolean; "v_value"?: number; "w"?: boolean; "w_value"?: number }
// ...
  "reused_cores__frame3d_linear__api__ElementInput": { "A": number; "Asy"?: (number) | (null); "Asz"?: (number) | (null); "E": number; "G": number; "Iy": number; "Iz": number; "J": number; "id": number; "node_i": number; "node_j": number; "reference_vector": [number, number, number]; "releases"?: Array<number>; "roll_angle"?: number; "theory"?: "euler_bernoulli" | "timoshenko" }
  "reused_cores__frame3d_linear__api__NodeInput": { "id": number; "x": number; "y": number; "z": number }
}
// ...
export type WithDefaults<T> = T extends object
  ? { [K in keyof T]-?: WithDefaults<T[K]> } : T